# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Teston HR Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /attendance:
    get:
      operationId: listAttendance
      tags: [attendance]
      summary: List attendance logs
      parameters:
        - in: query
          name: userId
          schema:
            type: number
        - in: query
          name: startDate
          schema:
            type: string
        - in: query
          name: endDate
          schema:
            type: string
        - in: query
          name: status
          schema:
            type: string
      responses:
        "200":
          description: Attendance logs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttendanceListResponse"
    post:
      operationId: createAttendance
      tags: [attendance]
      summary: Create attendance record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAttendanceRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttendanceResponse"
  /attendance/{id}:
    get:
      operationId: getAttendance
      tags: [attendance]
      summary: Get an attendance record
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Attendance
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttendanceResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateAttendance
      tags: [attendance]
      summary: Update attendance record
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAttendanceRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttendanceResponse"
  /attendance/checkin:
    post:
      operationId: checkIn
      tags: [attendance]
      summary: Check in with face data
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CheckInRequest"
      responses:
        "200":
          description: Checked in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttendanceResponse"
        "400":
          description: Already checked in or error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /attendance/checkout:
    post:
      operationId: checkOut
      tags: [attendance]
      summary: Check out
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CheckOutRequest"
      responses:
        "200":
          description: Checked out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttendanceResponse"
        "400":
          description: Not checked in or error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /attendance/summary:
    get:
      operationId: getAttendanceSummary
      tags: [attendance]
      summary: Get man-hours and cost summary
      parameters:
        - in: query
          name: startDate
          schema:
            type: string
        - in: query
          name: endDate
          schema:
            type: string
        - in: query
          name: departmentId
          schema:
            type: number
      responses:
        "200":
          description: Attendance summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttendanceSummaryResponse"
...
  /alerts:
    get:
      operationId: listAlerts
      tags: [alerts]
      summary: List alerts
      parameters:
        - in: query
          name: userId
          schema:
            type: number
        - in: query
          name: alertType
          schema:
            type: string
        - in: query
          name: status
          schema:
            type: string
        - in: query
          name: severity
          schema:
            type: string
      responses:
        "200":
          description: Alerts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AlertListResponse"
    post:
      operationId: createAlert
      tags: [alerts]
      summary: Create an alert
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAlertRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AlertResponse"
  /alerts/{id}:
    patch:
      operationId: updateAlert
      tags: [alerts]
      summary: Update alert status
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAlertRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AlertResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    LoginRequest:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required:
        - email
        - password
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/UserResponse"
      required:
        - token
        - user
    UserResponse:
      type: object
      properties:
        id:
          type: number
        name:
          type: string
        email:
          type: string
        role:
          type: string
        employeeId:
          type: string
          nullable: true
        departmentId:
          type: number
          nullable: true
        phone:
          type: string
          nullable: true
        position:
          type: string
          nullable: true
        hourlyRate:
          type: string
        isActive:
          type: string
        hasFaceRegistered:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - email
        - role
        - hourlyRate
        - isActive
        - hasFaceRegistered
        - createdAt
        - updatedAt
...
    AttendanceResponse:
      type: object
      properties:
        id:
          type: number
        userId:
          type: number
        checkIn:
          type: string
          format: date-time
          nullable: true
        checkOut:
          type: string
          format: date-time
          nullable: true
        date:
          type: string
        status:
          type: string
        hoursWorked:
          type: string
          nullable: true
        faceMatchScore:
          type: string
          nullable: true
        isManualEntry:
          type: string
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - date
        - status
        - isManualEntry
        - createdAt
        - updatedAt
    AttendanceListResponse:
      type: object
      properties:
        logs:
          type: array
          items:
            $ref: "#/components/schemas/AttendanceResponse"
        total:
          type: number
      required:
        - logs
        - total
    CreateAttendanceRequest:
      type: object
      properties:
        userId:
          type: number
        date:
          type: string
        status:
          type: string
          enum: [present, absent, late, half_day]
        checkIn:
          type: string
        checkOut:
          type: string
        notes:
          type: string
        isManualEntry:
          type: string
      required:
        - userId
        - date
    UpdateAttendanceRequest:
      type: object
      properties:
        status:
          type: string
          enum: [present, absent, late, half_day]
        checkIn:
          type: string
        checkOut:
          type: string
        notes:
          type: string
        hoursWorked:
          type: string
    CheckInRequest:
      type: object
      properties:
        userId:
          type: number
        faceDescriptor:
          type: string
          nullable: true
        faceMatchScore:
          type: string
          nullable: true
      required:
        - userId
    CheckOutRequest:
      type: object
      properties:
        userId:
          type: number
      required:
        - userId
    AttendanceSummaryItem:
      type: object
      properties:
        userId:
          type: number
        userName:
          type: string
        employeeId:
          type: string
          nullable: true
        department:
          type: string
          nullable: true
        totalDays:
          type: number
        presentDays:
          type: number
        absentDays:
          type: number
        lateDays:
          type: number
        totalHours:
          type: string
        totalCost:
          type: string
        hourlyRate:
          type: string
      required:
        - userId
        - userName
        - totalDays
        - presentDays
        - absentDays
        - lateDays
        - totalHours
        - totalCost
        - hourlyRate
    AttendanceSummaryResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/AttendanceSummaryItem"
        startDate:
          type: string
        endDate:
          type: string
        totalManHours:
          type: string
        totalCost:
          type: string
      required:
        - items
        - startDate
        - endDate
        - totalManHours
        - totalCost
...
    AlertResponse:
      type: object
      properties:
        id:
          type: number
        userId:
          type: number
          nullable: true
        alertType:
          type: string
        severity:
          type: string
        title:
          type: string
        message:
          type: string
        status:
          type: string
        relatedDate:
          type: string
          nullable: true
        resolvedBy:
          type: number
          nullable: true
        resolvedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - alertType
        - severity
        - title
        - message
        - status
        - createdAt
        - updatedAt
    AlertListResponse:
      type: object
      properties:
        alerts:
          type: array
          items:
            $ref: "#/components/schemas/AlertResponse"
        total:
          type: number
      required:
        - alerts
        - total
    CreateAlertRequest:
      type: object
      properties:
        userId:
          type: number
        alertType:
          type: string
          enum: [missing_attendance, missing_document, frequent_absence, policy_violation, ghost_worker, leave_limit_exceeded]
        severity:
          type: string
          enum: [info, warning, critical]
        title:
          type: string
        message:
          type: string
        relatedDate:
          type: string
      required:
        - alertType
        - title
        - message
    UpdateAlertRequest:
      type: object
      properties:
        status:
          type: string
          enum: [active, resolved, dismissed]
...