# lib/api-zod/src/generated/types/attendanceResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Teston HR Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface AttendanceResponse {
  id: number;
  userId: number;
  checkIn?: Date | null;
  checkOut?: Date | null;
  date: string;
  status: string;
  hoursWorked?: string | null;
  faceMatchScore?: string | null;
  isManualEntry: string;
  notes?: string | null;
  createdAt: Date;
  updatedAt: Date;
}