# lib/api-zod/src/generated/types/createAlertRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Teston HR Management System API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateAlertRequestAlertType } from "./createAlertRequestAlertType";
import type { CreateAlertRequestSeverity } from "./createAlertRequestSeverity";

export interface CreateAlertRequest {
  userId?: number;
  alertType: CreateAlertRequestAlertType;
  severity?: CreateAlertRequestSeverity;
  title: string;
  message: string;
  relatedDate?: string;
}