# lib/api-zod/src/generated/types/createAttendanceRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Teston HR Management System API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateAttendanceRequestStatus } from "./createAttendanceRequestStatus";

export interface CreateAttendanceRequest {
  userId: number;
  date: string;
  status?: CreateAttendanceRequestStatus;
  checkIn?: string;
  checkOut?: string;
  notes?: string;
  isManualEntry?: string;
}